update upgrade guide, move incrementingFileIterator back to utils
